requirements.txt
pyautogui
pytesseract
Pillow
pyperclip
pyinstaller
numpy<2